# perf: allow disabling forking on RQ worker