--- a/frappe/utils/background_jobs.py
+++ b/frappe/utils/background_jobs.py
@@ -197,7 +197,7 @@ def execute_job(site, method, event, job_name, kwargs, user=None, is_async=True,
 	retval = None
 
 	if is_async:
-		frappe.init(site=site)
+		frappe.init(site=site, force=True)
 		frappe.connect()
 		if os.environ.get("CI"):
 			frappe.flags.in_test = True