--- a/frappe/utils/background_jobs.py
+++ b/frappe/utils/background_jobs.py
@@ -16,7 +16,7 @@ from rq import Callback, Queue, Worker
 from rq.exceptions import NoSuchJobError
 from rq.job import Job, JobStatus
 from rq.logutils import setup_loghandlers
-from rq.worker import DequeueStrategy
+from rq.worker import DequeueStrategy, WorkerStatus
 from rq.worker_pool import WorkerPool
 from tenacity import retry, retry_if_exception_type, stop_after_attempt, wait_fixed
 
@@ -269,6 +269,22 @@ def execute_job(site, method, event, job_name, kwargs, user=None, is_async=True,
 			frappe.destroy()
 
 
+class FrappeWorker(Worker):
+	def __init__(self, *args, **kwargs):
+		self.disable_forking = kwargs.pop("disable_forking", None)
+		super().__init__(*args, **kwargs)
+
+	def execute_job(self, job: "Job", queue: "Queue"):
+		"""Execute job in same thread/process, do not fork()"""
+		if not self.disable_forking:
+			return super().execute_job(job, queue)
+
+		self.set_state(WorkerStatus.BUSY)
+		# TODO: death penalty stuff (?)
+		self.perform_job(job, queue)
+		self.set_state(WorkerStatus.IDLE)
+
+
 def start_worker(
 	queue: str | None = None,
 	quiet: bool = False,
@@ -302,7 +318,7 @@ def start_worker(
 	if quiet:
 		logging_level = "WARNING"
 
-	worker = Worker(queues, connection=redis_connection)
+	worker = FrappeWorker(queues, connection=redis_connection, disable_forking=True)
 	worker.work(
 		logging_level=logging_level,
 		burst=burst,