Basic overheads test 100 * `frappe.ping`

With forking
```
λ time bench worker --burst > /dev/null
bench worker --burst > /dev/null  1.28s user 1.33s system 91% cpu 2.870 total
```

Without forking:
```
λ time bench worker --burst > /dev/null
0.71s user 0.29s system 79% cpu 1.270 total
```


---

Somewhat realistic but empty job - 100 *  `frappe.email.queue.flush`

Before:
```
λ time bench worker --burst > /dev/null
12.71s user 7.27s system 98% cpu 20.387 total
```

After:
```
λ time bench worker --burst > /dev/null
0.45s user 0.21s system 77% cpu 0.861 total
```

Running all jobs in typical ERPNext install:
```py
for name in frappe.get_all("Scheduled Job Type", pluck="name"):
    frappe.get_doc("Scheduled Job Type", name).enqueue(True)

```

Before:
```
λ time bench worker --burst > /dev/null
23.67s user 12.46s system 96% cpu 37.500 total
```

After:
```
λ time bench worker --burst > /dev/null
5.72s user 2.33s system 88% cpu 9.136 total
```

---

TODO:
- [x] death penalty stuff, everything that RQ does for fork needs to be reimplemented
- [ ] This can cause memory leaks depending on how badly the job code is written, we should exit worker periodically (supervisor config restarts it)
- [ ] make it configurable