@ankush 2024-03-16
Cons:
- Need periodic restart - best suited for RQ Workerpool as it manges this
- More idle memory usage. Maybe stick to using this on large active benches with >50% utilization?